refactor(transformer, services): consolidate transformer logic into transformer.ts, remove deprecated transformer service, and enhance error handling for transformation steps
