+++ apps/app/src/lib/services/completion/index.ts
@@ -1,0 +1,17 @@
+import { AnthropicCompletionServiceLive } from './anthropic';
+import { GoogleCompletionServiceLive } from './google';
+import { GroqCompletionServiceLive } from './groq';
+import { OpenaiCompletionServiceLive } from './openai';
+
+export {
+	AnthropicCompletionServiceLive as anthropic,
+	GoogleCompletionServiceLive as google,
+	GroqCompletionServiceLive as groq,
+	OpenaiCompletionServiceLive as openai,
+};
+
+export type { AnthropicCompletionService } from './anthropic';
+export type { GoogleCompletionService } from './google';
+export type { GroqCompletionService } from './groq';
+export type { OpenaiCompletionService } from './openai';
+export type { CompletionService } from './_types';
