+++ apps/app/src/lib/services/index.ts
@@ -1,5 +1,6 @@
 import { SetTrayIconServiceLive } from './SetTrayIconService';
 import { ClipboardServiceLive } from './clipboard';
+import * as completions from './completion';
 import { CpalRecorderServiceLive } from './cpal-recorder';
 import { DbServiceLive } from './db';
 import { DownloadServiceLive } from './download';
@@ -10,7 +11,6 @@ import { NotificationServiceLive } from './notifications';
 import { OsServiceLive } from './os';
 import { PlaySoundServiceLive } from './sound';
 import * as transcriptions from './transcription';
-import { TransformerServiceLive } from './transformer';
 import { VadServiceLive } from './vad';
 
 /**
@@ -18,6 +18,7 @@ import { VadServiceLive } from './vad';
  */
 export {
 	ClipboardServiceLive as clipboard,
+	completions,
 	CpalRecorderServiceLive as cpalRecorder,
 	DbServiceLive as db,
 	DownloadServiceLive as download,
@@ -29,6 +30,5 @@ export {
 	SetTrayIconServiceLive as setTrayIcon,
 	PlaySoundServiceLive as sound,
 	transcriptions,
-	TransformerServiceLive as transformer,
 	VadServiceLive as vad,
 };
